refactor: rename ClusterSupport to RollingUpdateHandler
Extract the implementation into RollingUpdateHandler and deprecate
ClusterSupport as a thin subclass for backwards compatibility.
